pyproject.toml
[project]
name = "seven-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.10",
    "aiosignal==1.3.1",
    "alembic==1.14.0",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "argon2-cffi-bindings==21.2.0",
    "argon2-cffi==23.1.0",
    "async-timeout==4.0.3",
    "asyncpg==0.30.0",
    "attrs==24.2.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.12.3",
    "billiard==4.2.1",
    "boto3==1.35.76",
    "botocore==1.35.76",
    "cachetools==5.5.0",
    "celery==5.4.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.0",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "click==8.1.7",
    "cryptography==44.0.0",
    "dataclasses-json==0.6.7",
    "elastic-transport==8.15.1",
    "elasticsearch==8.16.0",
    "emoji==2.14.0",
    "eval-type-backport==0.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.6",
    "filetype==1.2.0",
    "frozenlist==1.5.0",
    "google-ai-generativelanguage==0.6.10",
    "google-api-core==2.23.0",
    "google-api-python-client==2.154.0",
    "google-auth-httplib2==0.2.0",
    "google-auth==2.36.0",
    "google-generativeai==0.8.3",
    "googleapis-common-protos==1.66.0",
    "greenlet==3.1.1",
    "grpcio-status==1.68.1",
    "grpcio==1.68.1",
    "h11==0.14.0",
    "html5lib==1.1",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx-sse==0.4.0",
    "httpx==0.28.1",
    "idna==3.10",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpath-python==1.0.6",
    "jsonpointer==3.0.0",
    "kombu==5.4.2",
    "langchain-community==0.3.10",
    "langchain-core==0.3.23",
    "langchain-google-genai==2.0.6",
    "langchain-text-splitters==0.3.2",
    "langchain==0.3.10",
    "langdetect==1.0.9",
    "langsmith==0.1.147",
    "loguru==0.7.3",
    "lxml==5.3.0",
    "mako==1.3.7",
    "markupsafe==3.0.2",
    "marshmallow==3.23.1",
    "minio==7.2.12",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "nltk==3.9.1",
    "numpy==1.26.4",
    "olefile==0.47",
    "orjson==3.10.12",
    "packaging==24.2",
    "prompt-toolkit==3.0.48",
    "propcache==0.2.1",
    "proto-plus==1.25.0",
    "protobuf==5.29.1",
    "psutil==6.1.0",
    "psycopg2-binary==2.9.9",
    "pyasn1-modules==0.4.1",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pycryptodome==3.21.0",
    "pydantic-core==2.23.4",
    "pydantic-settings==2.6.1",
    "pydantic==2.9.2",
    "pymupdf==1.25.0",
    "pyparsing==3.2.0",
    "pypdf==5.1.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-iso639==2024.10.22",
    "python-magic==0.4.27",
    "python-multipart==0.0.19",
    "python-oxmsg==0.0.1",
    "pyyaml==6.0.2",
    "rapidfuzz==3.10.1",
    "redis==5.2.1",
    "regex==2024.11.6",
    "requests-toolbelt==1.0.0",
    "requests==2.32.3",
    "rsa==4.9",
    "ruff==0.8.2",
    "s3transfer==0.10.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.36",
    "starlette==0.41.3",
    "tenacity==9.0.0",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2024.2",
    "unstructured-client==0.28.1",
    "unstructured==0.16.11",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
    "uvicorn==0.32.1",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "wrapt==1.17.0",
    "yarl==1.18.3",
    "jinja2==3.1.4",
    "pi-heif==0.21.0",
    "pillow==11.0.0",
    "pip==24.3.1",
    "pydantic-extra-types==2.10.1",
    "pyinstrument==5.0.0",
    "setuptools==75.6.0",
    "sqlparse==0.5.3",
    "iniconfig==2.0.0",
    "pluggy==1.5.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "tomli==2.2.1",
]
